Version: 1.1.0 Rename the interface file Separate Response and Request DTOs
