+++ src/main/java/com/openapi/gen/springboot/library_management/LibraryManagementApplication.java
@@ -1,13 +1,13 @@
-package com.openapi.gen.springboot.library_management_v1;
+package com.openapi.gen.springboot.library_management;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class LibraryManagementV1Application {
+public class LibraryManagementApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(LibraryManagementV1Application.class, args);
+		SpringApplication.run(LibraryManagementApplication.class, args);
 	}
 
 }
